requirements/test.txt
-r ./_common.txt

pytest
PyYAML
beautifulsoup4
